apy/pyproject.toml
[tool.poetry]
name = "apy"
version = "0.1.0"
description = ""
authors = ["Rohan Varma <rohan@adoptclarity.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.3"
gunicorn = "^22.0.0"
python-dotenv = "^1.0.1"
openai = "^1.58.1"
flask-cors = "^5.0.0"
typing-extensions = "^4.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
